+++ Vcare/app/Http/Controllers/Admin/bookingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Models\Admin;
 use App\Models\Doctor;
 use App\Models\Booking;
 use App\Models\Patient;
@@ -17,8 +18,10 @@ class BookingController extends Controller
      */
     public function index()
     {
+        $auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.view'), 403);
         $bookings = Booking::with('patient.user', 'doctor.user')->paginate(10);
-        return view('admin.pages.bookings.index', compact('bookings'));
+        return view('admin.pages.bookings.index', compact('bookings','auth'));
     }
 
     /**
@@ -26,6 +29,8 @@ class BookingController extends Controller
      */
     public function create()
     {
+        $auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.create'), 403);
         $doctors = Doctor::with('user')->get();
         $patients = Patient::with('user')->get();
         return view('admin.pages.bookings.create', compact('doctors', 'patients'));
@@ -35,7 +40,8 @@ class BookingController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(BookingRequest $request)
-    {
+    {$auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.create'), 403);
         $validated = $request->validated();
         Booking::create($validated);
         return redirect()->route('admin.bookings.index')->with('success', 'Booking created successfully.');
@@ -46,6 +52,8 @@ class BookingController extends Controller
      */
     public function show(string $id)
     {
+        $auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.view'), 403);
         $booking = Booking::with('patient.user', 'doctor.user')->findOrFail($id);
         return view('admin.pages.bookings.show', compact('booking'));
     }
@@ -55,7 +63,9 @@ class BookingController extends Controller
      */
     public function edit(Booking $booking)
     {
-        return view('admin.pages.bookings.edit', compact('booking'));
+        $auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.update'), 403);
+        return view('admin.pages.bookings.index', compact('booking'));
     }
 
     /**
@@ -63,12 +73,14 @@ class BookingController extends Controller
      */
     public function update(Request $request, Booking $booking)
     {
+        $auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.update'), 403);
         $validated = $request->validate([
             'status' => 'required|in:cancelled,visited,pending',
             'updated_at' => now(),
         ]);
         $booking->update($validated);
-        return redirect()->route('admin.bookings.index')->with('success', 'Booking updated successfully.');
+        return redirect()->back()->with('success', 'Booking updated successfully.');
     }
 
     /**
@@ -76,11 +88,15 @@ class BookingController extends Controller
      */
     public function destroy(Booking $booking)
     {
+        $auth =auth('admin')->user()->admin ?? auth()->user();
+        abort_if(!$auth || $auth->cannot('bookings.delete'), 403);
         $booking->delete();
         return redirect()->route('admin.bookings.index')->with('success', 'Booking deleted successfully.');
     }
     public function getDoctorAppointments($id)
     {
+        $auth = Admin::where('user_id', auth('admin')->id())->first() ?? abort(403, 'Unauthorized');
+        abort_if($auth->cannot('appointments.create'), 403);
         $appointments = Appointment::where('doctor_id', $id)->get();
         return response()->json($appointments);
     }
